mtsd_v2_fully_annotated/requirements.txt
Pillow>=4.3.0